Partition and ExtendableSquareMatrix source files/tests split into more files
