+++ tests/CMakeLists.txt
@@ -1,7 +1,8 @@
 add_executable(tests
 		ord-vector.cc
 		sparse-set.cc
-		partition-relation-pair.cc
+		partition.cc
+		extendable-square-matrix.cc
 		synchronized-iterator.cc
 		main.cc
 		alphabet.cc
